% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FifSampling.R
\name{double.condition.sampling}
\alias{double.condition.sampling}
\title{Divide a set of items into samples.}
\usage{
double.condition.sampling(data, feat1, feat2, sample.size)
}
\arguments{
\item{data}{a data frame, each line is an item.}

\item{feat1}{the column name containing the feature 1.}

\item{feat2}{the column name containing the feature 2.}

\item{sample.size}{size of the samples requested. The sample.size must be a multiple of the item number.}
}
\value{
A list of data frame.
}
\description{
Items are line of a data frame. Items have two properties: feat1 and feat2 (columns of the data frame).
The set of items is the cartesian product of feat1 and feat2.
The constraints on the sample are:
- Each item belongs to one and only one sample. The number of samples is then (nbr of items / sample size).
- Each sample contains only one representant of feat1. Then, the number of different feat1 cannot be lower than the number of sample, and the number of feat2 cannot be greater than the number of sample.
- No more than 2 consecutive identical feat1 modality in a sample.
- No more than 2 consecutive identical feat2 modality in a sample.
}
\examples{
sample.nb <- 5
data(dataFif)
samples <- double.condition.sampling(data=dataFif, "sent", "condition", 20)
}
